resolve assigned but unused
