+++ intake/conftest.py
@@ -132,3 +132,20 @@ def tempdir():
         yield d
     finally:
         shutil.rmtree(d)
+
+
+@pytest.fixture
+def temp_cache(tempdir):
+    import intake
+    old = intake.config.conf.copy()
+    olddir = intake.config.confdir
+    intake.config.confdir = str(tempdir)
+    intake.config.conf.update({'cache_dir': str(tempdir),
+                               'cache_download_progress': False,
+                               'cache_disabled': False})
+    intake.config.save_conf()
+    try:
+        yield
+    finally:
+        intake.config.confdir = olddir
+        intake.config.conf.update(old)
